requirements.txt
PyQt5==5.15.10
pypdf==4.1.0